Changed csXMLShaderTech::ActivatePass to match size_t prototype
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19095 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shadercompiler/xmlshader/xmlshader.cpp
@@ -586,7 +586,7 @@ bool csXMLShaderTech::Load (iDocumentNode* node, iDocumentNode* parentSV)
   return true;
 }
 
-bool csXMLShaderTech::ActivatePass (unsigned int number)
+bool csXMLShaderTech::ActivatePass (size_t number)
 {
   if(number>=passesCount)
     return false;
